LabelBase created and bitmap_label using it. Starting on label
