+++ adafruit_display_text/label.py
@@ -418,53 +418,3 @@ class Label(displayio.Group):
         self.height = self._font.get_bounding_box()[1]
         self._update_text(str(old_text))
         self.anchored_position = current_anchored_position
-
-    @property
-    def anchor_point(self):
-        """Point that anchored_position moves relative to.
-        Tuple with decimal percentage of width and height.
-        (E.g. (0,0) is top left, (1.0, 0.5): is middle right.)"""
-        return self._anchor_point
-
-    @anchor_point.setter
-    def anchor_point(self, new_anchor_point):
-        if self._anchor_point is not None:
-            current_anchored_position = self.anchored_position
-            self._anchor_point = new_anchor_point
-            self.anchored_position = current_anchored_position
-        else:
-            self._anchor_point = new_anchor_point
-
-    @property
-    def anchored_position(self):
-        """Position relative to the anchor_point. Tuple containing x,y
-        pixel coordinates."""
-        if self._anchor_point is None:
-            return None
-        return (
-            int(
-                self.x
-                + (self._boundingbox[0] * self.scale)
-                + round(self._anchor_point[0] * self._boundingbox[2] * self.scale)
-            ),
-            int(
-                self.y
-                + (self._boundingbox[1] * self.scale)
-                + round(self._anchor_point[1] * self._boundingbox[3] * self.scale)
-            ),
-        )
-
-    @anchored_position.setter
-    def anchored_position(self, new_position):
-        if (self._anchor_point is None) or (new_position is None):
-            return  # Note: anchor_point must be set before setting anchored_position
-        self.x = int(
-            new_position[0]
-            - (self._boundingbox[0] * self.scale)
-            - round(self._anchor_point[0] * (self._boundingbox[2] * self.scale))
-        )
-        self.y = int(
-            new_position[1]
-            - (self._boundingbox[1] * self.scale)
-            - round(self._anchor_point[1] * self._boundingbox[3] * self.scale)
-        )
